+++ lib/src/envy/data/data_group.dart
@@ -3,7 +3,7 @@ part of envy;
 /// A group node that also holds a map of named data sets.
 ///
 /// A named data set with the same name as one in an ancestor
-/// data node, including the global data pool will override
+/// data node, including the global data pool, will override
 /// the earlier data.
 ///
 class DataGroup extends GroupNode {