+++ lib/src/envy/multiplicity/largest_size.dart
@@ -4,10 +4,10 @@ class LargestSize extends Multiplicity {
   LargestSize();
 
   int sizeOf(Iterable<EnvyProperty> props) {
-    int largest = -1;
+    int largest = 0;
     for (var prop in props) {
-      //print("raw size = ${prop.rawSize}... ${prop}");
-      largest = Math.max(largest, prop.rawSize);
+      print("raw size = ${prop.rawSize}... ${prop}... optional ${prop.optional}... ${prop.payload}");
+      if (!prop.optional) largest = Math.max(largest, prop.rawSize);
     }
 
     return largest;