+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ### 0.0.7
-  * 
+  * Only update line dash value when necessary
   * Upgraded to `polymer 1.0.0-rc.14`, `polymer-elements 1.0.0-rc.7` and `reflectable 0.5.1`
 
 ### 0.0.6